refactor: refactor type map logic out to a new file
