+++ src/mongoose.repository.ts
@@ -2,7 +2,6 @@ import mongoose, {
   Connection,
   HydratedDocument,
   Model,
-  Schema,
   UpdateQuery,
 } from 'mongoose';
 import { Optional } from 'typescript-optional';
@@ -15,92 +14,7 @@ import {
   ValidationException,
 } from './util/exceptions';
 import { SaveOptions, SearchOptions } from './util/operation-options';
-
-/**
- * Models a domain object instance constructor.
- */
-export type Constructor<T> = new (...args: any) => T;
-
-/**
- * Models an abstract domain object supertype instance constructor.
- */
-export type AbsConstructor<T> = abstract new (...args: any) => T;
-
-/**
- * Models the data of a domain object subtype.
- */
-export type TypeData<T> = {
-  type: Constructor<T>;
-  schema: Schema;
-};
-
-/**
- * Models the data of a domain object supertype.
- */
-export type SupertypeData<T> = {
-  type: Constructor<T> | AbsConstructor<T>;
-  schema: Schema;
-};
-
-/**
- * Models a map of domain object subtypes supported by a custom repository.
- */
-export type SubtypeMap<T> = {
-  [key: string]: TypeData<T extends infer U ? U : never>;
-};
-
-/**
- * Models a map of domain object supertype and subtypes supported by a custom repository.
- */
-export type TypeMap<T extends Entity> =
-  | {
-      ['Default']: SupertypeData<T>;
-    }
-  | SubtypeMap<T>;
-
-class InnerTypeMap<T extends Entity> {
-  readonly types: string[];
-  readonly default: SupertypeData<T>;
-  readonly data: TypeData<T>[];
-
-  constructor(map: TypeMap<T>) {
-    if (!map.Default) {
-      throw new IllegalArgumentException(
-        'The given map must include domain supertype data',
-      );
-    }
-    this.default = map.Default as SupertypeData<T>;
-    this.types = Object.keys(map).filter((key) => key !== 'Default');
-    this.data = Object.entries(map).reduce((accumulator, entry) => {
-      if (entry[0] !== 'Default') {
-        // @ts-expect-error - safe instantiation as any non-root map entry refers to some subtype data
-        accumulator.push(entry[1]);
-      }
-      return accumulator;
-    }, []);
-  }
-
-  getSubtypeData(type: string): TypeData<T> | undefined {
-    const index = this.types.indexOf(type);
-    return index !== -1 ? this.data[index] : undefined;
-  }
-
-  getSupertypeData(): TypeData<T> | SupertypeData<T> {
-    return this.default;
-  }
-
-  getSupertypeName(): string {
-    return this.getSupertypeData().type.name;
-  }
-
-  getSubtypesData(): TypeData<T>[] {
-    return this.data;
-  }
-
-  has(type: string): boolean {
-    return type === this.getSupertypeName() || this.types.indexOf(type) !== -1;
-  }
-}
+import { Constructor, TypeMap, TypeMapImpl } from './util/type-map';
 
 /**
  * Abstract Mongoose-based implementation of the {@link Repository} interface.
@@ -108,7 +22,7 @@ class InnerTypeMap<T extends Entity> {
 export abstract class MongooseRepository<T extends Entity & UpdateQuery<T>>
   implements Repository<T>
 {
-  private readonly typeMap: InnerTypeMap<T>;
+  private readonly typeMap: TypeMapImpl<T>;
   protected readonly entityModel: Model<T>;
 
   /**
@@ -120,7 +34,7 @@ export abstract class MongooseRepository<T extends Entity & UpdateQuery<T>>
     typeMap: TypeMap<T>,
     protected readonly connection?: Connection,
   ) {
-    this.typeMap = new InnerTypeMap(typeMap);
+    this.typeMap = new TypeMapImpl(typeMap);
     this.entityModel = this.createEntityModel(connection);
   }
